Controller: Iso Packaging
This is the controller job to create the ISO Package and upload it to
an AWS S3 endpoint.

This commits needs commits from PR #28, where it adds the iso_package
command to the platform.

Related to: https://issues.redhat.com/browse/ECOPROJECT-767
Related to: https://issues.redhat.com/browse/ECOPROJECT-831

Signed-off-by: Eloy Coto <eloy.coto@acalustra.com>